@@ -8,20 +8,21 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@react-three/fiber": "^7.0.26",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "three": "^0.138.3"
+    "@react-three/drei": "^9.4.2",
+    "@react-three/fiber": "^8.0.11",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "three": "^0.139.2"
   },
   "devDependencies": {
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
-    "@types/three": "^0.138.0",
-    "@vitejs/plugin-react": "^1.0.7",
+    "@types/react": "^18.0.2",
+    "@types/react-dom": "^18.0.0",
+    "@types/three": "^0.139.0",
+    "@vitejs/plugin-react": "^1.3.0",
     "autoprefixer": "^10.4.4",
     "postcss": "^8.4.12",
     "tailwindcss": "^3.0.23",
-    "typescript": "^4.5.4",
-    "vite": "^2.8.0"
+    "typescript": "^4.6.3",
+    "vite": "^2.9.1"
   }
 }
